# Removed Hopac dependency