--- a/src/FSharp.Data.GraphQL.Server/paket.references
+++ b/src/FSharp.Data.GraphQL.Server/paket.references
@@ -1,3 +1,2 @@
 FSharp.Core
-Hopac
 FSharp.Quotations.Evaluator