--- a/src/FSharp.Data.GraphQL.Shared/FSharp.Data.GraphQL.Shared.fsproj
+++ b/src/FSharp.Data.GraphQL.Shared/FSharp.Data.GraphQL.Shared.fsproj
@@ -40,6 +40,7 @@
     <Compile Include="Extensions.fs" />
     <Compile Include="Prolog.fs" />
     <Compile Include="Ast.fs" />
+    <Compile Include="AsyncVal.fs" />
     <Compile Include="TypeSystem.fs" />
     <Compile Include="Validation.fs" />
     <Compile Include="Introspection.fs" />