… tests/FSharp.Data.GraphQL.Benchmarks/AsyncValBenchmark.fs
… tests/FSharp.Data.GraphQL.Benchmarks/Program.fs
… tests/FSharp.Data.GraphQL.Tests.Sql/FSharp.Data.GraphQL.Tests.Sql.fsproj
… tests/FSharp.Data.GraphQL.Tests/AsyncValTests.fs
… tests/FSharp.Data.GraphQL.Tests/FSharp.Data.GraphQL.Tests.fsproj
… tests/FSharp.Data.GraphQL.Tests/paket.references